feat: moved components to Algorithm components folder
