+++ src/components/Algorithms/contents/Definition.tsx
