+++ src/components/Algorithms/contents/Measurement.tsx
