+++ src/components/Algorithms/contents/QuickSort.tsx
