+++ src/components/Algorithms/contents/index.ts
