+++ src/components/Algorithms/index.ts
@@ -1,0 +1,4 @@
+export * from './contents'
+export { default as ContentContainer } from './ContentContainer'
+export { default as Section } from './Section'
+export { default as SideMenu } from './SideMenu'
